Added css and js files, little changed files architecture
